Add virtul hosted-style & subresource signed URLs
We should be encouraging users to switch to virtual hosted-style signed
URLs, as this will be the preferred style in the nearish future. This
keeps the default to path-style (as it was before), but allows
specifying "virtual_hosted" via an optional CLI flag.

This also adds support for specifying a subresource, via passing in the
query parameter corresponding to the subresource you want to access. To
get a bucket or object's ACLs, you'd specify the "acl" query parameter
in the XML API endpoint -- this change allows you to specify it as a CLI
flag, e.g. "--subresource=acl".